Back-end API working on the remote server
* finally got the API working on the remote server. It looks like a lot of
changes but these are just comments and refactoring.
Important change is the method definition in Check class from "get"
to "post"
* In terms of extension added new method to handle sending POST requests
* It looks like previously I had most of the code written as if it was
POST request, but I was using GET.
Don't know how this worked locally, but it's better that way.
And anyway in requirements I've stated to use POST.
Moreover, if I used GET it has a limit on length of URL.
So in theory if the URL requested was near that limit it would break.